Add instruction to access self object
It turns out the compiler was doing some very signficant optimizations
based on the fast that the only ScriptObjects that could be created -
Array and Dictionary - didn't return RunError. Having to account for
this now results in quite a significant slowdown.

It may be worth looking into returning only `CallError`s and forgoing
the `RunError` type (i.e. add a `RunError` variant to `CallError`).